ZKL now shows level + elo in tooltip of each user (to test new protocol extensions)
svn path=/trunk/other/; revision=3993

+++ NightWatch/AuthService.cs
@@ -48,7 +48,7 @@ namespace NightWatch
               if (acc != null) {
                   client.Extensions.Publish(e.Data.Name, new Dictionary<string, string>() { 
                   {ProtocolExtension.Keys.Level.ToString() , acc.Level.ToString()},
-                      {ProtocolExtension.Keys.EffectiveElo.ToString(), acc.EffectiveElo.ToString()}
+                      {ProtocolExtension.Keys.EffectiveElo.ToString(), ((int)acc.EffectiveElo).ToString()}
                   
                   });
